Added pre and post create random data delegates
The final method generateData() wraps around the abstract data method in
order to raise delegates when calling it. Two delegates,
EntriesPreCreateRandomData and EntriesPostCreateRandomData are fired
before and after the call to data().

This commit also include the changes needed in order for the save
content page to use the new generateData() method.