Refactor of random entries for associations
There was a couple of problems with the way randomness was used, which
always resulted in poor results. In order to be able to have good
distribution of values, we need to do somes check and filter out some
information at different levels.

Hence, the current change makes sure that all related fields in select
box fields are gonna check if the data table exists and also check that
there are data in them: this removes the problem of having no data found
in a required field, where there are choices in on relation over two
possible fields.

It also adds a random check to sometime remove the section where there
is only one possible value.

Fixes #7
Fixes #10